requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
requests==2.32.3
httpx==0.28.1
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.2
pyyaml==6.0.2
